# Conflicts:
#	LoRaEngine/LoraKeysManagerFacade/HttpUtilities.cs
#	LoRaEngine/LoraKeysManagerFacade/SearchDeviceByDevEUI.cs
#	LoRaEngine/LoraKeysManagerFacade/VersionValidator.cs
#	LoRaEngine/modules/LoRaWanNetworkSrvModule/LoRaWan.NetworkServer/ServiceFacadeHttpClientProvider.cs
#	LoRaEngine/modules/LoRaWanNetworkSrvModule/LoraTools/CommonAPI/ApiVersion.cs
#	LoRaEngine/test/LoRaWanNetworkServer.Test/ServiceFacadeHttpClientHandlerTest.cs
#	LoRaEngine/test/LoraKeysManagerFacade.Test/ApiValidationTest.cs
#	LoRaEngine/test/LoraKeysManagerFacade.Test/ApiVersionTest.cs
#	LoRaEngine/test/LoraKeysManagerFacade.Test/HttpUtilitiesTest.cs
#	LoRaEngine/test/LoraKeysManagerFacade.Test/SearchDeviceByDevEUITest.cs
#	Tests/Integration/DevAddrCacheTest.cs
#	Tests/Integration/FunctionBundlerTests.cs
#	Tests/Shared/LoRaWan.Tests.Shared.csproj
#	Tests/Shared/WaitableLoRaRequest.cs

+++ LoRaEngine/test/LoRaWanNetworkServer.Test/MultiGatewayFrameCounterUpdateStrategyTest.cs
@@ -143,7 +143,7 @@ namespace LoRaWan.NetworkServer.Test
             this.deviceApi.Setup(x => x.NextFCntDownAsync(device.DevEUI, It.IsAny<uint>(), It.IsAny<uint>(), this.gatewayID))
                 .Returns<string, uint, uint, string>((devEUI, fcntDown, payloadFcnt, gatewayID) =>
                 {
-                    return Task.FromResult((uint)(fcntDown + 1));
+                    return Task.FromResult(fcntDown + 1);
                 });
 
             for (var i = 1; i <= 15; i++)
